requirements.txt
autopep8==1.4.3
boto3==1.20.14
botocore==1.23.14
boxsdk==3.7.2
coverage
Django==4.2.11
django-cors-headers==3.11.0
django-filter==21.1
django-mysql==4.12.0
django-oauth-toolkit==1.7.1
django-redis==5.2.0
djangorestframework==3.14.0
drf-writable-nested==0.6.3
docutils==0.14
factory-boy==2.12.0
google-cloud-vision==0.37.0
gunicorn==19.9.0
jmespath==0.9.3
mysqlclient==2.2.4
numpy==1.23.0
opencv-python==4.6.0.66
pandas==1.4.3
protobuf==3.20.1
ptvsd==4.2.3
pycodestyle==2.5.0
python-dateutil==2.8.2
pytz==2022.2.1
pyyaml==6.0
pytest-freezegun==0.4.2
pillow==9.1.1
pdf2image==1.12.1
s3transfer==0.5
six==1.16.0
sqlalchemy==1.3.6 # pandasのdataframe.to_sqlを使用するため
sendgrid==6.9.7
tblib==1.4.0
urllib3==1.26.13
PyDrive2==1.6.2
google-api-python-client==1.12.1
requests==2.27.1
payjp==0.1.0
Jinja2==2.11.1
MarkupSafe==2.0.1
pdfkit==0.6.1
mojimoji==0.0.11 # 全角<->半角の変換を行うため
dropbox==10.4.1
cryptography==3.4.6
pypdf==3.12.0
Pycryptodome==3.18.0
sentry-sdk==0.20.3
PyJWT==2.4.0
faker==9.3.1
aws-xray-sdk==2.8.0
pdfservices-sdk==2.1.2 # 通常PDF変換エラー時のPDF変換
aiohttp==3.8.1
openpyxl==3.0.9
tqdm
moto==4.2.14
typing-extensions==4.12.1
zeep==4.2.1
more-itertools==9.0.0
chardet==4.0.0
newrelic
jsondiff==2.0.0
charset-normalizer==2.0.12
tnefparse==1.4.0
PyMuPDF==1.24.4
